feat: Add vendor phone number setting
Introduces a new `vendorPhoneNumber` setting to `SystemSettings`. This allows vendors to configure a phone number that will be displayed and used for click-to-call functionality in the Customer PWA. The initial default value is set for backward compatibility.

Also includes minor refactors to the Vendor App's `OrderGrid` and `SubViewWrapper` to improve component structure.